reorganize layout
