+++ x-pack/platform/plugins/shared/alerting_v2/public/components/rule_details/rule_detail_page.tsx
@@ -6,11 +6,9 @@
  */
 
 import {
-  EuiBadge,
   EuiButtonEmpty,
   EuiCodeBlock,
   EuiDescriptionList,
-  EuiFlexGroup,
   EuiFlexItem,
   EuiPageHeader,
   EuiPanel,
@@ -28,6 +26,7 @@ import { RuleDetailsActionsMenu } from './rule_details_actions_menu';
 import { useBreadcrumbs } from '../../hooks/use_breadcrumbs';
 import { useDeleteRule } from '../../hooks/use_delete_rule';
 import { RulesDeleteModalConfirmation } from '../common/rules_delete_modal_confirmation';
+import { RuleHeaderDescription } from './rule_header_description';
 
 export interface RuleDetailPageProps {
   rule: RuleApiResponse;
@@ -68,32 +67,6 @@ const formatMaybeList = (value: string[] | undefined) => {
   }
 
   return value.join(', ');
-};
-
-const formatMaybeIsoDateTime = (value: string | undefined) => {
-  if (!value) {
-    return EMPTY_VALUE;
-  }
-
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    return EMPTY_VALUE;
-  }
-
-  const datePart = new Intl.DateTimeFormat(undefined, {
-    year: 'numeric',
-    month: 'short',
-    day: '2-digit',
-  }).format(date);
-  const timePart = new Intl.DateTimeFormat(undefined, {
-    hour: 'numeric',
-    minute: '2-digit',
-  }).format(date);
-
-  return i18n.translate('xpack.alertingV2.ruleDetails.dateTime', {
-    defaultMessage: '{date} at {time}',
-    values: { date: datePart, time: timePart },
-  });
 };
 
 export const RuleDetailPage: React.FunctionComponent<RuleDetailPageProps> = ({ rule }) => {
@@ -164,45 +137,6 @@ export const RuleDetailPage: React.FunctionComponent<RuleDetailPageProps> = ({ r
     [rule.schedule, rule.time_field, rule.grouping?.fields]
   );
 
-  const metadataListItems = useMemo(
-    () => [
-      {
-        title: i18n.translate('xpack.alertingV2.ruleDetails.created', {
-          defaultMessage: 'Created',
-        }),
-        description: (
-          <ItemValueRuleSummary
-            data-test-subj="alertingV2RuleDetailsCreatedAt"
-            itemValue={formatMaybeIsoDateTime(rule.createdAt)}
-          />
-        ),
-      },
-      {
-        title: i18n.translate('xpack.alertingV2.ruleDetails.lastModified', {
-          defaultMessage: 'Last modified',
-        }),
-        description: (
-          <ItemValueRuleSummary
-            data-test-subj="alertingV2RuleDetailsUpdatedAt"
-            itemValue={formatMaybeIsoDateTime(rule.updatedAt)}
-          />
-        ),
-      },
-      {
-        title: i18n.translate('xpack.alertingV2.ruleDetails.createdBy', {
-          defaultMessage: 'Created by',
-        }),
-        description: (
-          <ItemValueRuleSummary
-            data-test-subj="alertingV2RuleDetailsCreatedBy"
-            itemValue={formatMaybeString(rule.createdBy)}
-          />
-        ),
-      },
-    ],
-    [rule.createdAt, rule.updatedAt, rule.createdBy]
-  );
-
   return (
     <>
       {showDeleteConfirmation && (
@@ -223,83 +157,7 @@ export const RuleDetailPage: React.FunctionComponent<RuleDetailPageProps> = ({ r
             />
           </span>
         }
-        description={
-          <EuiFlexGroup
-            alignItems="center"
-            justifyContent="flexStart"
-            gutterSize="m"
-            wrap={false}
-            responsive={false}
-          >
-            <EuiFlexItem grow={false}>
-              <EuiFlexGroup responsive={false} alignItems="center">
-                <EuiFlexItem grow={false}>
-                  {rule.enabled ? (
-                    <EuiBadge color="success" data-test-subj="enabledBadge">
-                      <FormattedMessage
-                        id="xpack.triggersActionsUI.sections.ruleDetails.enabled"
-                        defaultMessage="Enabled"
-                      />
-                    </EuiBadge>
-                  ) : (
-                    <EuiBadge color="subdued" data-test-subj="disabledBadge">
-                      <FormattedMessage
-                        id="xpack.triggersActionsUI.sections.ruleDetails.disabled"
-                        defaultMessage="Disabled"
-                      />
-                    </EuiBadge>
-                  )}
-                </EuiFlexItem>
-              </EuiFlexGroup>
-            </EuiFlexItem>
-            <EuiFlexItem grow={false}>
-              <EuiFlexGroup responsive={false} gutterSize="s" alignItems="center">
-                <EuiFlexItem grow={false}>
-                  <EuiText size="xs" color="subdued" data-test-subj="ruleType">
-                    <FormattedMessage
-                      id="xpack.triggersActionsUI.sections.ruleDetails.kindLabel"
-                      defaultMessage="Kind"
-                    />
-                  </EuiText>
-                </EuiFlexItem>
-                <EuiFlexItem grow={false}>
-                  <EuiBadge color="hollow" data-test-subj="kindBadge">
-                    <FormattedMessage
-                      id="xpack.triggersActionsUI.sections.ruleDetails.kindValue"
-                      defaultMessage="{ruleType}"
-                      values={{ ruleType: rule.kind }}
-                    />
-                  </EuiBadge>
-                </EuiFlexItem>
-              </EuiFlexGroup>
-            </EuiFlexItem>
-            <EuiFlexItem grow={false}>
-              <EuiFlexGroup responsive={false} gutterSize="s" alignItems="center">
-                <EuiFlexItem grow={false}>
-                  <EuiText size="xs" color="subdued" data-test-subj="ruleTags">
-                    <FormattedMessage
-                      id="xpack.triggersActionsUI.sections.ruleDetails.tagsLabel"
-                      defaultMessage="Tags"
-                    />
-                  </EuiText>
-                </EuiFlexItem>
-                <EuiFlexItem grow={false}>
-                  {rule.metadata.labels ? (
-                    <EuiFlexGroup gutterSize="xs" wrap responsive={false} data-test-subj="ruleTags">
-                      {rule.metadata.labels.map((label) => (
-                        <EuiFlexItem key={label} grow={false}>
-                          <EuiBadge color="hollow">{label}</EuiBadge>
-                        </EuiFlexItem>
-                      ))}
-                    </EuiFlexGroup>
-                  ) : (
-                    '-'
-                  )}
-                </EuiFlexItem>
-              </EuiFlexGroup>
-            </EuiFlexItem>
-          </EuiFlexGroup>
-        }
+        description={<RuleHeaderDescription rule={rule} />}
         rightSideItems={[
           <RuleDetailsActionsMenu
             key="actions"
@@ -327,66 +185,47 @@ export const RuleDetailPage: React.FunctionComponent<RuleDetailPageProps> = ({ r
       />
 
       <EuiSpacer size="m" />
-      <EuiFlexGroup responsive={false}>
-        <EuiFlexItem grow={2}>
-          <EuiPanel color="subdued" hasBorder={false} paddingSize="m">
-            <EuiTitle size="s">
-              <EuiFlexItem grow={false}>
-                {i18n.translate('xpack.triggersActionsUI.ruleDetails.definition', {
-                  defaultMessage: 'Configuration',
-                })}
-              </EuiFlexItem>
-            </EuiTitle>
-            <EuiSpacer size="m" />
-            <EuiDescriptionList
-              compressed={true}
-              type="column"
-              listItems={configurationListItems}
-              css={{ alignItems: 'start' }}
-            />
 
-            <EuiSpacer size="m" />
-            <EuiTitle size="xs">
-              <h3>
-                {i18n.translate('xpack.alertingV2.ruleDetails.query', {
-                  defaultMessage: 'ES|QL query',
-                })}
-              </h3>
-            </EuiTitle>
-            <EuiSpacer size="s" />
-            <EuiPanel hasBorder paddingSize="none" css={{ width: '100%' }}>
-              <EuiCodeBlock
-                language="text"
-                isCopyable
-                overflowHeight={360}
-                paddingSize="m"
-                data-test-subj="alertingV2RuleDetailsQuery"
-                css={{ width: '100%' }}
-              >
-                {rule.evaluation?.query?.base || EMPTY_VALUE}
-              </EuiCodeBlock>
-            </EuiPanel>
-          </EuiPanel>
-        </EuiFlexItem>
-        <EuiFlexItem grow={1}>
-          <EuiPanel hasBorder paddingSize="m">
-            <EuiTitle size="s">
-              <EuiFlexItem grow={false}>
-                {i18n.translate('xpack.alertingV2.ruleDetails.metadata', {
-                  defaultMessage: 'Metadata',
-                })}
-              </EuiFlexItem>
-            </EuiTitle>
-            <EuiSpacer size="m" />
-            <EuiDescriptionList
-              compressed={true}
-              type="column"
-              listItems={metadataListItems}
-              css={{ alignItems: 'start', gridTemplateColumns: '140px 1fr' }}
-            />
-          </EuiPanel>
-        </EuiFlexItem>
-      </EuiFlexGroup>
+      <EuiPanel color="subdued" hasBorder={false} paddingSize="m">
+        <EuiTitle size="s">
+          <EuiText>
+            {i18n.translate('xpack.triggersActionsUI.ruleDetails.definition', {
+              defaultMessage: 'Configuration',
+            })}
+          </EuiText>
+        </EuiTitle>
+        <EuiSpacer size="m" />
+        <EuiDescriptionList
+          compressed={true}
+          type="column"
+          listItems={configurationListItems}
+          css={{ alignItems: 'start' }}
+        />
+      </EuiPanel>
+      <EuiSpacer size="s" />
+
+      <EuiPanel color="subdued" hasBorder={false} paddingSize="m">
+        <EuiTitle size="s">
+          <EuiText>
+            {i18n.translate('xpack.triggersActionsUI.ruleDetails.query', {
+              defaultMessage: 'ES|QL Query',
+            })}
+          </EuiText>
+        </EuiTitle>
+        <EuiSpacer size="s" />
+        <EuiPanel hasBorder paddingSize="none" css={{ width: '100%' }}>
+          <EuiCodeBlock
+            language="esql"
+            isCopyable
+            overflowHeight={360}
+            paddingSize="m"
+            data-test-subj="alertingV2RuleDetailsQuery"
+            css={{ width: '100%' }}
+          >
+            {rule.evaluation?.query?.base || EMPTY_VALUE}
+          </EuiCodeBlock>
+        </EuiPanel>
+      </EuiPanel>
 
       <EuiSpacer size="m" />
     </>
